Calculate library PCR concentrations

Calculate library PCR concentrations

Usage

pcr_lib_calc(pcr, dil_factor = 1000)

Arguments

pcr

a pcr object. Will be tidied if not already.

dil_factor

integer. The factor that the libraries were diluted for pcr

Value

a pcr object, containing the input columns as well as:

standard_diff The differences between the ct_mean of a standard and one step up in the dilution (ie more concentrated, lower Ct). The most concentrated dilution has a value of 0

dil 2^(standard_diff). The accuracy of this metric assumes that the efficiency of the PCR is 100%, which is likely good but not perfect! In the case of the first standard, dil = 0

quant_actual For standards, the presumed quantity of standard, calculated from dil. For samples, quantity

concentrationThe concentration of library, before dilution

Examples

system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  read_pcr() |>
  pcr_lib_calc()
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   .default = col_double(),
#>   `Well Position` = col_character(),
#>   Omit = col_logical(),
#>   `Sample Name` = col_character(),
#>   `Target Name` = col_character(),
#>   Task = col_character(),
#>   Reporter = col_character(),
#>   Quencher = col_character(),
#>   CT = col_character(),
#>   `Automatic Ct Threshold` = col_logical(),
#>   `Automatic Baseline` = col_logical(),
#>   Comments = col_logical()
#> )
#> ℹ Use `spec()` for the full column specifications.
#> ! Multiple files in zip: reading ''[Content_Types].xml''
#> pcr is not tidy, tidying.
#> <pcr[8]>
#> 
#>                 12
#>    ________________________
#>   | ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯
#>   | ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯
#>   | ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯
#>   | ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯
#> 8 | ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯
#>   | ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯
#>   | ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯
#>   | ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯ ◯
#> 
#> 
#> Start corner: tl
#> Plate dimensions: 8 x 12
#> # Experiment type: standard_curve 
#> # Wells: 96 
#> # Is tidy: TRUE 
#> # Date: 2020-03-11