Create library prep quality control data

Create library prep quality control data

pcr_lib_qc(lib_calc_pcr)

Arguments

lib_calc_pcr
an output from pcr_lib_calc

Value

a named list with:

standards Data for individual standards, including calculated dilutions, given and calculated quantities, raw Ct, etc.

samples Data for individual samples, including calculated concentrations, raw Ct, etc.

sample_summary Summary statistics for samples grouped by replicates

standard_summary Summary statistics for standards groupd by replicates

outliers Data for individual samples and standards with and without their putative outliers (po) per replicate group

Details

While the output of this function on its own is can theoretically be used to gauge library quality, it is best used in conjunction with a function like pcr_lib_calc_report

Examples

system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  pcr_tidy(pad_zero = TRUE) |>
  pcr_lib_calc() |>
  pcr_lib_qc()

#> $standards
#> # A tibble: 13 x 8
#> # Groups:   task [1]
#>    task     quantity quant_actual   dil slope efficiency    r2    ct
#>    <chr>       <dbl>        <dbl> <dbl> <dbl>      <dbl> <dbl> <dbl>
#>  1 STANDARD 6.80         6.8       0    -3.48       93.7 0.999  15.4
#>  2 STANDARD 6.80         6.8       0    -3.48       93.7 0.999  15.5
#>  3 STANDARD 6.80         6.8       0    -3.48       93.7 0.999  15.5
#>  4 STANDARD 0.680        0.734     9.27 -3.48       93.7 0.999  18.7
#>  5 STANDARD 0.680        0.734     9.27 -3.48       93.7 0.999  18.6
#>  6 STANDARD 0.0680       0.0603   12.2  -3.48       93.7 0.999  22.3
#>  7 STANDARD 0.0680       0.0603   12.2  -3.48       93.7 0.999  22.2
#>  8 STANDARD 0.0680       0.0603   12.2  -3.48       93.7 0.999  22.3
#>  9 STANDARD 0.00680      0.00568  10.6  -3.48       93.7 0.999  25.7
#> 10 STANDARD 0.00680      0.00568  10.6  -3.48       93.7 0.999  25.6
#> 11 STANDARD 0.00680      0.00568  10.6  -3.48       93.7 0.999  25.7
#> 12 STANDARD 0.000680     0.000417 13.6  -3.48       93.7 0.999  29.4
#> 13 STANDARD 0.000680     0.000417 13.6  -3.48       93.7 0.999  29.4
#> 
#> $samples
#> # A tibble: 42 x 10
#> # Groups:   task [1]
#>    task    sample_name quantity_mean concentration quantity quant_actual slope
#>    <chr>   <chr>               <dbl>         <dbl>    <dbl>        <dbl> <dbl>
#>  1 UNKNOWN Sample 06            2.04         2039.     2.06         2.06 -3.48
#>  2 UNKNOWN Sample 06            2.04         2039.     2.12         2.12 -3.48
#>  3 UNKNOWN Sample 06            2.04         2039.     1.94         1.94 -3.48
#>  4 UNKNOWN Sample 12            1.89         1893.     1.93         1.93 -3.48
#>  5 UNKNOWN Sample 12            1.89         1893.     1.88         1.88 -3.48
#>  6 UNKNOWN Sample 12            1.89         1893.     1.87         1.87 -3.48
#>  7 UNKNOWN Sample 04            1.69         1694.     1.62         1.62 -3.48
#>  8 UNKNOWN Sample 04            1.69         1694.     1.84         1.84 -3.48
#>  9 UNKNOWN Sample 04            1.69         1694.     1.62         1.62 -3.48
#> 10 UNKNOWN Sample 16            1.49         1493.     1.46         1.46 -3.48
#> # ... with 32 more rows, and 3 more variables: efficiency <dbl>, r2 <dbl>,
#> #   ct <dbl>
#> 
#> $sample_summary
#> # A tibble: 16 x 3
#>    sample_name quantity_mean concentration_mean
#>    <chr>               <dbl>              <dbl>
#>  1 Sample 01           0.599               599.
#>  2 Sample 02           1.05               1052.
#>  3 Sample 03           1.37               1367.
#>  4 Sample 04           1.69               1694.
#>  5 Sample 05           0.903               903.
#>  6 Sample 06           2.04               2039.
#>  7 Sample 07           1.13               1132.
#>  8 Sample 08           1.09               1088.
#>  9 Sample 09           1.22               1218.
#> 10 Sample 10           1.07               1066.
#> 11 Sample 11           1.03               1031.
#> 12 Sample 12           1.89               1893.
#> 13 Sample 13           0.583               583.
#> 14 Sample 14           1.22               1216.
#> 15 Sample 15           0.713               713.
#> 16 Sample 16           1.49               1493.
#> 
#> $standard_summary
#> # A tibble: 10 x 4
#>    quantity   dil name             value
#>       <dbl> <dbl> <chr>            <dbl>
#>  1 0.000680 13.6  quantity_mean 0.000680
#>  2 0.000680 13.6  quant_actual  0.000417
#>  3 0.00680  10.6  quantity_mean 0.00680 
#>  4 0.00680  10.6  quant_actual  0.00568 
#>  5 0.0680   12.2  quantity_mean 0.0680  
#>  6 0.0680   12.2  quant_actual  0.0603  
#>  7 0.680     9.27 quantity_mean 0.680   
#>  8 0.680     9.27 quant_actual  0.734   
#>  9 6.80      0    quantity_mean 6.80    
#> 10 6.80      0    quant_actual  6.8     
#> 
#> $outliers
#> # A tibble: 58 x 17
#> # Groups:   sample_name [22]
#>    task     sample_name quantity_mean   dil slope efficiency    r2    ct quantity
#>    <chr>    <chr>               <dbl> <dbl> <dbl>      <dbl> <dbl> <dbl>    <dbl>
#>  1 STANDARD 1                   NA        0 -3.48       93.7 0.999  15.4     6.80
#>  2 STANDARD 1                   NA        0 -3.48       93.7 0.999  15.5     6.80
#>  3 STANDARD 1                   NA        0 -3.48       93.7 0.999  15.5     6.80
#>  4 UNKNOWN  Sample 06            2.04    NA -3.48       93.7 0.999  17.1     2.06
#>  5 UNKNOWN  Sample 06            2.04    NA -3.48       93.7 0.999  17.1     2.12
#>  6 UNKNOWN  Sample 06            2.04    NA -3.48       93.7 0.999  17.2     1.94
#>  7 UNKNOWN  Sample 12            1.89    NA -3.48       93.7 0.999  17.2     1.93
#>  8 UNKNOWN  Sample 12            1.89    NA -3.48       93.7 0.999  17.3     1.88
#>  9 UNKNOWN  Sample 12            1.89    NA -3.48       93.7 0.999  17.3     1.87
#> 10 UNKNOWN  Sample 04            1.69    NA -3.48       93.7 0.999  17.5     1.62
#> # ... with 48 more rows, and 8 more variables: quant_actual <dbl>,
#> #   concentration <dbl>, no_po_mean <dbl>, no_po_sd <dbl>, keep <lgl>,
#> #   adj_mean <dbl>, adj_sd <dbl>, z <dbl>
#>