Plan PCR experiment

Plan PCR experiment

Usage

pcr_plan(
  data,
  n_primers,
  format = 384,
  exclude_border = TRUE,
  primer_names = NULL,
  headless = TRUE,
  has_names = TRUE
)

Arguments

data

a data.frame, with samples as the first column (if has_names = TRUE) and RNA concentrations as the second (or first, if has_names = FALSE)

n_primers

integer. Number of primers to be used in the experiment.

format

integer. 96 or 384 - the number of wells of the plate planned to be used

exclude_border

logical. Should the border be excluded to avoid edge effects? Default is TRUE.

primer_names

character vector. Names of primers.

headless

logical. If FALSE, return invisible and redirect to shiny application.

has_names

logical. Is the first column the names of the samples?

Value

a named list

Examples

dummy_rna_conc |>
  pcr_plan(n_primers = 3)
#> $mm_prep
#> # A tibble: 4 × 2
#>   reagent             vol
#>   <chr>             <dbl>
#> 1 2X RT-PCR Buffer  206. 
#> 2 Primer             20.6
#> 3 25X RT-PCR Enzyme  16.5
#> 4 Nuclease Free H2O 103. 
#> 
#> $sample_prep
#> # A tibble: 8 × 7
#>   sample       conc dilution_factor diluted_concentration diluted_rna_to_add
#>   <chr>       <dbl>           <int>                 <dbl>              <dbl>
#> 1 24hr_DMSO_1 127.                1                 127.                1.18
#> 2 48hr_DMSO_1  93                 1                  93                 1.61
#> 3 24hr_1uM_1  143.                1                 143.                1.05
#> 4 48hr_1uM_1   67.1               1                  67.1               2.24
#> 5 24hr_DMSO_2  88.3               1                  88.3               1.70
#> 6 48hr_DMSO_2 124.                1                 124.                1.21
#> 7 24hr_1uM_2   94.2               1                  94.2               1.59
#> 8 48hr_1uM_2   80.2               1                  80.2               1.87
#> # ℹ 2 more variables: water_to_add <dbl>, final_vol <dbl>
#> 
#> $plate
#> 
#>       3
#>    ______
#> 1 | ◯ ◯ ◯
#> 
#> 
#> Start corner: tl
#> Plate dimensions: 16 x 24
#> 
#> $n_primers
#> [1] 3
#> 
#> $format
#> [1] "384"
#> 
#> $exclude_border
#> [1] TRUE
#> 
#> $primer_names
#> NULL
#>