Create a report from a PCR plan

Create a report from a PCR plan

pcr_plan_report(pcr_plan, file_path = NULL)

Arguments

pcr_plan
output from pcr_plan

file_path
Where the report should be written, as well as the file name. Defaults to temp file.

Value

a named list, like the output of pcr_plan, but with the output file path appended.

Examples

dummy_rna_conc |>
  pcr_plan(n_primers = 3) |>
  pcr_plan_report()

#> Warning: cannot create dir '', reason 'No such file or directory'

#> Warning: path[1]="": The filename, directory name, or volume label syntax is incorrect

#> 
#> 
#> processing file: pcr_report-template.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.........                                                             |  12%
#>    inline R code fragments
#> 
#>   |                                                                              |..................                                                    |  25%
#> label: unnamed-chunk-1 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                              |..........................                                            |  38%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................................                                   |  50%
#> label: unnamed-chunk-2 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                              |............................................                          |  62%
#>   ordinary text without R code
#> 
#>   |                                                                              |....................................................                  |  75%
#> label: unnamed-chunk-3 (with options) 
#> List of 4
#>  $ echo      : logi FALSE
#>  $ fig.width : num 10
#>  $ fig.height: num 7
#>  $ out.width : chr "100%"
#> 

#> Warning: cannot create dir '\\2021-10-15_pcr-report_ccc20f9cdd_files\figure-html', reason 'Invalid argument'

#> Warning: unable to open file '//2021-10-15_pcr-report_ccc20f9cdd_files/figure-html/unnamed-chunk-3-1.png' for writing

#> Warning: opening device failed

#> Quitting from lines 99-112 (pcr_report-template.Rmd) 

#> Error in png(..., res = dpi, units = "in"): unable to start png() device

#>