Create library prep quality control data

Create library prep quality control data

Usage

pcr_lib_qc(lib_calc_pcr)

Arguments

lib_calc_pcr

A pcr object, output from pcr_lib_calc

Value

a pcr object with list with:

standards Data for individual standards, including calculated dilutions, given and calculated quantities, raw Ct, etc.

samples Data for individual samples, including calculated concentrations, raw Ct, etc.

sample_summary Summary statistics for samples grouped by replicates

standard_summary Summary statistics for standards groupd by replicates

outliers Data for individual samples and standards with and without their putative outliers (po) per replicate group

Details

While the output of this function on its own is can theoretically be used to gauge library quality, it is best used in conjunction with a function like pcr_lib_calc_report

Examples

system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  pcr_tidy(pad_zero = TRUE) |>
  pcr_lib_calc() |>
  pcr_lib_qc()
#> Error in pcr_tidy(system.file("extdata", "untidy-standard-curve.xlsx",     package = "amplify"), pad_zero = TRUE): could not find function "pcr_tidy"